api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
python-dotenv
pydantic-settings
